fix some review comments
